Add lib/repositories.js to handle merging all repositories together
