+++ lib/voting/voting.js
@@ -32,7 +32,7 @@
       }
 
 
-      var Voting = function Voting(repo, votingConfig) {
+      var Voting = function Voting(repo) {
         var self = this;
 
         this.starGazers = {};
@@ -41,12 +41,12 @@
         this.votesPerPR = {};
 
         this.repo = repo;
-        this.votingConfig = votingConfig;
+        this.votingConfig = repo.votingConfig;
         this.votingConfig.truePeriod =
-          this.votingConfig.period * (1 + (Math.random() - 0.5) * votingConfig.period_jitter);
+          this.votingConfig.period * (1 + (Math.random() - 0.5) * this.votingConfig.period_jitter);
         this.votingConfig.guaranteedResult =
           Math.ceil(this.votingConfig.minVotes * this.votingConfig.supermajority);
-        this.voteController = new VoteController(repo, votingConfig, this.starGazers, this.votesPerPR);
+        this.voteController = new VoteController(repo, this.votingConfig, this.starGazers, this.votesPerPR);
         this.templates = templates(this.votingConfig);
 
         // Only initialize the regular polling if the bot is not being tested.
